refactor: integrate canvas rendering logic into MediaCanvas component, remove useCanvasRenderer hook
