+++ packages/vscode-extension/src/webview/Preview/MediaCanvas.tsx
@@ -1,12 +1,18 @@
-import { useEffect, forwardRef, RefObject, useRef } from "react";
+import { useEffect, forwardRef, RefObject, useRef, useMemo, useCallback } from "react";
 import { useProject } from "../providers/ProjectProvider";
-import useCanvasRenderer from "../hooks/useCanvasRenderer";
+import { DeviceRotation } from "../../common/Project";
 
 type MediaRef = RefObject<HTMLImageElement | null> | RefObject<HTMLVideoElement | null>;
 
 interface MediaCanvasProps extends React.CanvasHTMLAttributes<HTMLCanvasElement> {
   mediaRef: MediaRef;
   src?: string;
+}
+
+interface TransformationConfig {
+  angle: number;
+  isPortrait: boolean;
+  rotation: DeviceRotation;
 }
 
 const MediaCanvas = forwardRef<HTMLCanvasElement, MediaCanvasProps>(
@@ -16,14 +22,102 @@ const MediaCanvas = forwardRef<HTMLCanvasElement, MediaCanvasProps>(
 
     const isRunningRef = useRef(false);
     const canvasRef = ref as RefObject<HTMLCanvasElement>;
-    const drawToCanvas = useCanvasRenderer(rotation, canvasRef);
+    const lastCanvasDimensionsRef = useRef<{ width: number; height: number } | null>(null);
+
+    // Memoize the transformation configuration based on rotation
+    const transformConfig = useMemo<TransformationConfig>(() => {
+      switch (rotation) {
+        case DeviceRotation.LandscapeLeft:
+          return {
+            angle: -Math.PI / 2,
+            isPortrait: false,
+            rotation: DeviceRotation.LandscapeLeft,
+          };
+        case DeviceRotation.LandscapeRight:
+          return {
+            angle: Math.PI / 2,
+            isPortrait: false,
+            rotation: DeviceRotation.LandscapeRight,
+          };
+        case DeviceRotation.PortraitUpsideDown:
+          return {
+            angle: Math.PI,
+            isPortrait: true,
+            rotation: DeviceRotation.PortraitUpsideDown,
+          };
+        default:
+          return {
+            angle: 0,
+            isPortrait: true,
+            rotation: DeviceRotation.Portrait,
+          };
+      }
+    }, [rotation]);
+
+    const drawToCanvas = useCallback(
+      (sourceImg: HTMLImageElement | HTMLVideoElement): void => {
+        if (!canvasRef.current) {
+          return;
+        }
+        const canvas = canvasRef.current;
+        const ctx = canvas.getContext("2d");
+        if (!ctx) {
+          return;
+        }
+
+        const sourceWidth =
+          sourceImg instanceof HTMLVideoElement ? sourceImg.videoWidth : sourceImg.width;
+        const sourceHeight =
+          sourceImg instanceof HTMLVideoElement ? sourceImg.videoHeight : sourceImg.height;
+
+        // Early return if image dimensions are invalid
+        if (sourceWidth === 0 || sourceHeight === 0) {
+          return;
+        }
+
+        // Calculate actual dimensions based on rotation
+        const newWidth = transformConfig.isPortrait ? sourceWidth : sourceHeight;
+        const newHeight = transformConfig.isPortrait ? sourceHeight : sourceWidth;
+
+        const currentCanvasDims = { width: newWidth, height: newHeight };
+
+        const canvasDimsChanged =
+          !lastCanvasDimensionsRef.current ||
+          lastCanvasDimensionsRef.current.width !== newWidth ||
+          lastCanvasDimensionsRef.current.height !== newHeight;
+
+        if (canvasDimsChanged) {
+          canvas.width = newWidth;
+          canvas.height = newHeight;
+          lastCanvasDimensionsRef.current = currentCanvasDims;
+        }
+
+        // Clear canvas
+        ctx.clearRect(0, 0, newWidth, newHeight);
+
+        if (transformConfig.rotation === DeviceRotation.Portrait) {
+          // Direct draw for portrait mode
+          ctx.drawImage(sourceImg, 0, 0);
+        } else {
+          // Apply transformation from transformConfig
+          ctx.save();
+          ctx.translate(newWidth / 2, newHeight / 2);
+          ctx.rotate(transformConfig.angle);
+          ctx.translate(-sourceWidth / 2, -sourceHeight / 2);
+          ctx.drawImage(sourceImg, 0, 0);
+          ctx.restore();
+        }
+      },
+      [canvasRef, transformConfig]
+    );
+
     const drawToCanvasRef = useRef(drawToCanvas);
 
     useEffect(() => {
       drawToCanvasRef.current = drawToCanvas;
     }, [drawToCanvas]);
 
-    // Unified canvas rendering effect for both image and video
+    // Unified canvas rendering for both image and video
     useEffect(() => {
       const mediaElement = mediaRef.current;
       const canvas = canvasRef.current;
@@ -78,7 +172,7 @@ const MediaCanvas = forwardRef<HTMLCanvasElement, MediaCanvasProps>(
         video.addEventListener("error", handleSourceError);
       } else {
         throw new Error(
-          "Unsupported media type for MediaCanvas. Only HTMLImageElement (MJPEG) and HTMLVideoElement are supported."
+          "Unsupported media type for MediaCanvas. Only HTMLImageElement (MJPEG src) and HTMLVideoElement are supported."
         );
       }
